+++ lib/flutter_wk.dart
@@ -5,36 +5,39 @@ import 'package:flutter/services.dart';
 /// Integrate a Widget into your App.
 /// {@endtemplate}
 class WidgetKit {
+  /// {@macro flutter_wk}
+  const WidgetKit._();
+
+  /// The method channel used to interact with the native platform.
   static const MethodChannel _channel = MethodChannel('ziqq.github/flutter_wk');
 
   /// Reloads all timelines for all widgets.
-  static Future<void> reloadAllTimelines() =>
-      _channel.invokeMethod<void>('reloadAllTimelines');
+  static Future<void> reload() => _channel.invokeMethod<void>('reload');
 
   /// Reloads timelines for widgets of a specific kind.
-  static Future<void> reloadTimelines(String ofKind) =>
-      _channel.invokeMethod<void>('reloadTimelines', <String, Object?>{
-        'ofKind': ofKind,
-      });
+  static Future<void> reloadOfKind(String kind) => _channel.invokeMethod<void>(
+    'reloadOfKind',
+    <String, Object?>{'ofKind': kind},
+  );
 
-  /// Gets an item from the shared app group storage.
-  static Future<T?> getItem<T>(String key, String appGroup) =>
-      _channel.invokeMethod<T>('getItem', <String, Object?>{
+  /// Reads an item from the shared app group storage.
+  static Future<T?> read<T>(String key, String appGroup) =>
+      _channel.invokeMethod<T>('read', <String, Object?>{
         'key': key,
         'appGroup': appGroup,
       });
 
-  /// Sets an item in the shared app group storage.
-  static Future<T?> setItem<T>(String key, Object? value, String appGroup) =>
-      _channel.invokeMethod<T>('setItem', <String, Object?>{
+  /// Writes an item to the shared app group storage.
+  static Future<T?> write<T>(String key, Object? value, String appGroup) =>
+      _channel.invokeMethod<T>('write', <String, Object?>{
         'key': key,
         'value': value,
         'appGroup': appGroup,
       });
 
   /// Removes an item from the shared app group storage.
-  static Future<bool> removeItem(String key, String appGroup) async =>
-      await _channel.invokeMethod<bool>('removeItem', <String, Object?>{
+  static Future<bool> remove(String key, String appGroup) async =>
+      await _channel.invokeMethod<bool>('remove', <String, Object?>{
         'key': key,
         'appGroup': appGroup,
       }) ??